requirements.txt
mysql-connector-python==8.3.0
tk==0.1.0
reportlab==4.1.0